adding models and stuff
